Update and rename DG-LAB-V3-SOCKET-To-V2-BLE-V1.0.3R.py to DG-LAB-V3-SOCKET-To-V2-BLE-V1.0.0semver.py
增加了个性化功能（包含调整界面强调色、修改背景图片）
增加了手动控制未连接时不可用的限制
增加了一张默认背景图
增加了代码注释提高开发者可读性
增加了 Semantic 规范版本号
修改了部分翻译
移除了浅色模式
修复了已知问题
